Week 6 beveiligd met datatype check einde les woensdag 13-12-2017

+++ delete.php
@@ -1,15 +1,23 @@
 <?php
-  // We gaan contact maken in de code met de mysql database
-  include("./connect_db.php");
+  // Voorkom een sql-injection bij delete  
+  if (is_numeric($_GET["id"])) {
+    // We gaan contact maken in de code met de mysql database
+    include("./connect_db.php");
+  
+    // We maken een select query
+    $sql = "DELETE FROM `users` 
+    WHERE `id` = " . $_GET["id"];
 
-  // We maken een select query
-  $sql = "DELETE FROM `users` 
-                WHERE `id` = " . $_GET["id"];
+    // echo $sql; exit();
+    // Vuur de sql-query af op de mysql-database
+    $result = mysqli_query($conn,$sql);
 
-  // echo $sql; exit();
-  // Vuur de sql-query af op de mysql-database
-  $result = mysqli_query($conn,$sql);
+    // Leidt de gebruiker terug naar de index.php pagina
+    header("Location: ./index.php?action=succes");
+  }
+  
 
+  
   // Leidt de gebruiker terug naar de index.php pagina
-  header("Location: ./index.php");
+  header("Location: ./index.php?action=fail");
 ?>
